# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NOFX Lite API
  version: 1.0.0
  description: |
    NOFX Lite 后端 API 文档（基于 OpenAPI v3.0.3）。

    本服务是单进程 Express 应用，默认监听 `http://127.0.0.1:3100`，所有业务接口
    前缀均为 `/api`。前端（Vue 3 + Vite）开发模式下通过 5173 → 3100 代理转发。

    约定：
      * 所有请求与响应均为 `application/json`，UTF-8 编码。
      * 通用错误响应体：`{ "error": "人类可读中文消息" }`，HTTP 状态码随错误类型变化。
      * 长耗时接口（`/api/market/analyze*`、`/api/research/performance/refresh`、
        `/api/history/fetch`）单次运行可达数分钟，前端应按长轮询/进度条轮询处理。
      * 写操作的请求体采用「合并语义」：未传入的字段不会清空既有配置（参见 `store.js#mergeConfig`）。
      * 配置中的密钥字段（`binance.apiKey/secretKey`、`okx.apiKey/secretKey/passphrase`、
        `model.apiKey`）始终以 `xxxx...xxxx` 形式回显，原值在 `maskConfig` 中屏蔽。
servers:
  - url: http://127.0.0.1:3100
    description: 本地开发/生产（PM2 单进程）
...
paths:
...
  /api/analyses:
    get:
      tags: [Analyses]
      summary: 研究快照列表（按日期/币种过滤，分页）
      parameters:
        - in: query
          name: date
          schema: { type: string, description: 'YYYY-MM-DD' }
        - in: query
          name: symbol
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: 快照数组
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AnalysisRecord' }
  /api/analyses/{id}:
    get:
      tags: [Analyses]
      summary: 单条研究快照详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 快照详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisRecord' }
        '404':
          description: 未找到
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: '未找到该分析记录。' }

  # ───────── Market analyze (research) ─────────
...
  /api/research/performance:
    get:
      tags: [Research]
      summary: 信号回测聚合（按策略/币种/方向分组）
      description: |
        取最近 500 条分析记录，对每条 signal 用存档 K 线回放，
        模拟从 `firstEntryAt` 到当前（或超时）的资金曲线。
      parameters:
        - in: query
          name: date
          schema: { type: string }
        - in: query
          name: symbol
          schema: { type: string }
      responses:
        '200':
          description: 回测结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PerformanceResult' }
  /api/research/performance/refresh:
    post:
      tags: [Research]
      summary: 触发一次"取最新 K 线 + 重算"（长耗时）
      description: |
        与 `GET /api/research/performance` 等价，但会按当前 `marketData` 行情源
        补齐缺失窗口的 K 线。受同一时刻只能有一个 refresh 任务的并发锁保护。
      responses:
        '200':
          description: 重算结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PerformanceResult' }
        '409':
          description: 行情正在同步
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: '模拟行情正在同步，请稍后重试。' }

  # ───────── Paper ─────────
...
  /api/paper/automation/{kind}:
    post:
      tags: [PaperAutomation]
      deprecated: true
      summary: Replaced by the two global automation tasks
      parameters:
        - in: path
          name: kind
          required: true
          schema: { type: string }
      responses:
        '410':
          description: Use /api/automation/tasks/{taskName}/trigger
...
components:
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: '人类可读中文消息' }
    JsonObject:
      type: object
      additionalProperties: true

    # ───── Config ─────
...
    AnalysisSignal:
      type: object
      properties:
        symbol: { type: string }
        interval: { type: string }
        eligible: { type: boolean }
        confidence: { type: number, minimum: 0, maximum: 1 }
        confidenceType: { type: string, example: rule_strength }
        analysisEngine: { type: string, enum: [local, local-mtf, ai] }
        positionRecommendation: { type: string, enum: [OPEN_LONG, OPEN_SHORT, HOLD] }
        firstEntryAt: { type: string, format: date-time }
        reasons:
          type: array
          items: { type: string }
        risks:
          type: array
          items: { type: string }
        invalidation: { type: string }
        marketProvider: { type: string, enum: [okx, binance] }
    AnalysisRecord:
      type: object
      properties:
        id: { type: string }
        at: { type: string, format: date-time }
        type: { type: string, enum: [single, range, all] }
        scope: { $ref: '#/components/schemas/AnalyzeScope' }
        analysisEngine: { type: string }
        analyses:
          type: array
          items: { $ref: '#/components/schemas/AnalysisSignal' }
        error: { type: string, nullable: true }
    AnalyzeScope:
      type: object
      properties:
        interval: { type: string, example: 15m }
        limit: { type: integer, minimum: 20, maximum: 200, example: 80 }
        maxSymbols: { type: integer, minimum: 1, maximum: 300, example: 20 }
        batchSize: { type: integer, minimum: 1, maximum: 20, example: 10 }
        engine: { type: string, enum: [auto, local, local-mtf, ai] }
        symbols:
          type: array
          items: { type: string }
        symbolsText: { type: string }
    PerformanceResult:
      type: object
      properties:
        asOf: { type: string, format: date-time }
        records: { type: integer }
        truncated: { type: boolean }
        excluded: { type: integer }
        summary: { $ref: '#/components/schemas/PerformanceSummary' }
        byStrategy:
          type: array
          items: { $ref: '#/components/schemas/PerformanceSummary' }
        bySymbol:
          type: array
          items: { $ref: '#/components/schemas/PerformanceSummary' }
        byDirection:
          type: array
          items: { $ref: '#/components/schemas/PerformanceSummary' }
        items:
          type: array
          items: { $ref: '#/components/schemas/PerformanceItem' }
        errors:
          type: array
          items: { type: string }
    PerformanceSummary:
      type: object
      properties:
        key: { type: string }
        count: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
        winRate: { type: number, minimum: 0, maximum: 1 }
        net: { type: number }
        avgNet: { type: number }
    PerformanceItem:
      type: object
      properties:
        id: { type: string }
        recordId: { type: string }
        symbol: { type: string }
        interval: { type: string }
        direction: { type: string, enum: [OPEN_LONG, OPEN_SHORT, HOLD] }
        strategyVersion: { type: string }
        at: { type: string, format: date-time }
        confidence: { type: number }
        costs: { $ref: '#/components/schemas/JsonObject' }
        evaluation: { $ref: '#/components/schemas/JsonObject' }

    # ───── Paper ─────
...